+++ mockspresso-mockito/src/main/java/com/episode6/hackit/mockspresso/mockito/MockitoMockMaker.java
@@ -6,7 +6,12 @@ import org.mockito.Mockito;
 
 /**
  * A {@link com.episode6.hackit.mockspresso.api.MockerConfig.MockMaker} for Mockito
+ *
+ * @deprecated This functionality is internal and should not be exposed.
+ *
+ * This class will be marked internal/protected in a future release
  */
+@Deprecated
 public class MockitoMockMaker implements MockerConfig.MockMaker {
 
   @SuppressWarnings("unchecked")